requirements.txt
fastapi>=0.89.1
uvicorn>=0.20.0
FlagEmbedding>=1.3.4
gradio>=3.50.2
pyjwt>=2.10.1
python-dotenv>=0.21.1
requests>=2.31.0 